requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
pydot==2.0.0
pyparsing==3.1.1
